Split refer contexts into their own file
