+++ src/index.ts
@@ -14,6 +14,10 @@ export { LoggerFactory } from "./LoggerFactory";
 export { NameAddrHeader } from "./NameAddrHeader";
 export { Parser } from "./Parser";
 export { PublishContext } from "./PublishContext";
+export {
+  ReferClientContext,
+  ReferServerContext
+} from "./ReferContext";
 export { RegisterContext } from "./RegisterContext";
 export { RequestSender } from "./RequestSender";
 
@@ -25,8 +29,6 @@ export { ServerContext } from "./ServerContext";
 export {
   InviteClientContext,
   InviteServerContext,
-  ReferClientContext,
-  ReferServerContext,
   Session
 } from "./Session";
 export {
